+++ src/storage/db/mod.rs
@@ -2,4 +2,6 @@ pub use self::rocksdb::*;
 
 mod multi_chunk_writer;
 mod rocksdb;
+
+pub mod backup;
 pub mod snapshot;
